+++ fastapi_jsonapi/schema_builder.py
@@ -7,7 +7,6 @@ from typing import (
     Iterable,
     List,
     Optional,
-    Set,
     Tuple,
     Type,
     TypeVar,
@@ -15,10 +14,9 @@ from typing import (
 )
 
 import pydantic
-from pydantic import BaseConfig, root_validator, validator
+from pydantic import BaseConfig
 from pydantic import BaseModel as PydanticBaseModel
-from pydantic.class_validators import ROOT_VALIDATOR_CONFIG_KEY, VALIDATOR_CONFIG_KEY
-from pydantic.fields import FieldInfo, ModelField, Validator
+from pydantic.fields import FieldInfo, ModelField
 
 from fastapi_jsonapi.data_typing import TypeSchema
 from fastapi_jsonapi.schema import (
@@ -35,6 +33,10 @@ from fastapi_jsonapi.schema import (
 )
 from fastapi_jsonapi.schema_base import BaseModel, Field, RelationshipInfo, registry
 from fastapi_jsonapi.splitter import SPLIT_REL
+from fastapi_jsonapi.validation_utils import (
+    extract_field_validators,
+    extract_validators,
+)
 
 JSON_API_RESPONSE_TYPE = Dict[Union[int, str], Dict[str, Any]]
 
@@ -291,7 +293,10 @@ class SchemaBuilder:
                 # works both for to-one and to-many
                 included_schemas.append((name, field.type_, relationship.resource_type))
             elif name == "id":
-                id_validators = self._extract_field_validators(schema, target_field_name="id")
+                id_validators = extract_field_validators(
+                    schema,
+                    include_for_field_names={"id"},
+                )
                 resource_id_field = (*(resource_id_field[:-1]), id_validators)
 
                 if not field.field_info.extra.get("client_can_set_id"):
@@ -310,7 +315,7 @@ class SchemaBuilder:
             f"{base_name}AttributesJSONAPI",
             **attributes_schema_fields,
             __config__=ConfigOrmMode,
-            __validators__=self._extract_validators(schema, exclude_for_field_names={"id"}),
+            __validators__=extract_validators(schema, exclude_for_field_names={"id"}),
         )
 
         relationships_schema = pydantic.create_model(
@@ -377,111 +382,6 @@ class SchemaBuilder:
         )
         self.relationship_schema_cache[cache_key] = relationship_data_schema
         return relationship_data_schema
-
-    def _is_target_validator(self, attr_name: str, value: Any, validator_config_key: str) -> bool:
-        """
-        True if passed object is validator of type identified by "validator_config_key" arg
-
-        :param attr_name:
-        :param value:
-        :param validator_config_key: Choice field, available options are pydantic consts
-                                     VALIDATOR_CONFIG_KEY, ROOT_VALIDATOR_CONFIG_KEY
-        """
-        return (
-            # also with private items
-            not attr_name.startswith("__")
-            and getattr(value, validator_config_key, None)
-        )
-
-    def _unpack_validators(self, model: Type[BaseModel], validator_config_key: str) -> Dict[str, Validator]:
-        """
-        Selects all validators from model attrs and unpack them from class methods
-
-        :param model: Type[BaseModel]
-        :param validator_config_key: Choice field, available options are pydantic consts
-                                     VALIDATOR_CONFIG_KEY, ROOT_VALIDATOR_CONFIG_KEY
-        """
-        root_validator_class_methods = {
-            # validators only
-            attr_name: value
-            for attr_name, value in model.__dict__.items()
-            if self._is_target_validator(attr_name, value, validator_config_key)
-        }
-
-        return {
-            validator_name: getattr(validator_method, validator_config_key)
-            for validator_name, validator_method in root_validator_class_methods.items()
-        }
-
-    def _extract_root_validators(self, model: Type[BaseModel]) -> Dict[str, Callable]:
-        validators = {}
-
-        unpacked_validators = self._unpack_validators(model, ROOT_VALIDATOR_CONFIG_KEY)
-        for validator_name, validator_instance in unpacked_validators.items():
-            validators[validator_name] = root_validator(
-                pre=validator_instance.pre,
-                skip_on_failure=validator_instance.skip_on_failure,
-                allow_reuse=True,
-            )(validator_instance.func)
-
-        return validators
-
-    def _extract_field_validators(
-        self,
-        model: Type[BaseModel],
-        target_field_name: str = None,
-        exclude_for_field_names: Set[str] = None,
-    ) -> Dict[str, Callable]:
-        """
-        :param model: Type[BaseModel]
-        :param target_field_name: Name of field for which validators will be returned.
-                                  If not set the function will return validators for all fields.
-        """
-        validators = {}
-        validator_origin_param_keys = ("pre", "each_item", "always", "check_fields")
-
-        unpacked_validators = self._unpack_validators(model, VALIDATOR_CONFIG_KEY)
-        for validator_name, (field_names, validator_instance) in unpacked_validators.items():
-            if target_field_name and target_field_name not in field_names:
-                continue
-            elif target_field_name:
-                field_names = [target_field_name]  # noqa: PLW2901
-
-            if exclude_for_field_names:
-                field_names = [  # noqa: PLW2901
-                    # filter names
-                    field_name
-                    for field_name in field_names
-                    if field_name not in exclude_for_field_names
-                ]
-
-            if not field_names:
-                continue
-
-            validators[validator_name] = validator(
-                *field_names,
-                allow_reuse=True,
-                **{
-                    # copy origin params
-                    param_name: getattr(validator_instance, param_name)
-                    for param_name in validator_origin_param_keys
-                },
-            )(validator_instance.func)
-
-        return validators
-
-    def _extract_validators(
-        self,
-        model: Type[BaseModel],
-        exclude_for_field_names: Set[str] = None,
-    ) -> Dict[str, Callable]:
-        return {
-            **self._extract_field_validators(
-                model,
-                exclude_for_field_names=exclude_for_field_names,
-            ),
-            **self._extract_root_validators(model),
-        }
 
     def _build_jsonapi_object(
         self,